wip repro issue with get compatible netlists
